Started on the Navigation bars on the top and bottom. Added pages for the bottom tabs. Still working on the top bar
