+++ app/(tabs)/gymfinder.tsx
@@ -1,5 +1,5 @@
 export default function TabTwoScreen() {
   return (
-    <p>Character</p>
+    <p>Gym Finder</p>
   );
 }
